moving plot bands test to new file
